Fix underwater swimming calc due to assigning collision height to uint32

+++ src/game/Entities/Player.cpp
@@ -18902,7 +18902,7 @@ void Player::SetOriginalGroup(Group* group, int8 subgroup)
 
 void Player::UpdateTerainEnvironmentFlags(Map* m, float x, float y, float z)
 {
-    uint32 collisionHeight = GetCollisionHeight();
+    float collisionHeight = GetCollisionHeight();
     GridMapLiquidData liquid_status;
     GridMapLiquidStatus res = m->GetTerrain()->getLiquidStatus(x, y, z, MAP_ALL_LIQUIDS, &liquid_status, collisionHeight);
     if (!res)
